+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # Release notes
 
+## New in git-machete 2.14.0
+
+- fixed: format of logs and other messages
+
 ## New in git-machete 2.13.6
 
 - fixed: remove reviewer setting when opening a PR to NixOS/nixpkgs